add_product_via_barcode/requirements.txt
python-stdnum==0.9